L2MTaskDetector

Documentation for L2MTaskDetector.

L2MTaskDetector.TaskDetectorMetrics

L2MTaskDetector.TaskDetectorMetrics

L2MTaskDetector.TaskDetectorModule

L2MTaskDetector.TaskDetectorModule

L2MTaskDetector.TaskDetectorParameters

L2MTaskDetector.TaskDetectorParameters

L2MTaskDetector.classify

L2MTaskDetector.sigmoid

L2MTaskDetector.sigmoid

L2MTaskDetector.task_feature_preprocess

L2MTaskDetector.train!

L2MTaskDetector.update_metrics

L2MTaskDetector.window_average

L2MTaskDetector.TaskDetectorMetrics — Type

TaskDetectorMetrics

Stateful information of the C3's metrics.

Fields

cvis::Vector{AbstractCVI}: list of CVIs used in computing metrics for C3.

values::Vector{Vector{Float64}}: most recent CVI criterion values.

performance::Float64: Meta-ICVI value.

n_window::Int64: number of values to track for Meta-ICVI calculation.

L2MTaskDetector.TaskDetectorMetrics — Method

TaskDetectorMetrics()

Default constructor for the C3 metrics object.

L2MTaskDetector.TaskDetectorModule — Type

TaskDetectorModule

The stateful information of C3.

Fields

ddvfa::DDVFA: the first processing step of C3.

metrics::TaskDetectorMetrics: metrics container for C3, containing ICVIs and more.

task_map::Dict{String, Int}: internal mapping of string task names to incremental indices.

task::L2MTask: the most recent task belief of C3.

params::TaskDetectorParameters: all algorithmic parameters.

L2MTaskDetector.TaskDetectorModule — Method

TaskDetectorModule(conf::ConfParse)

Default constructor for the C3 stateful information, requires an L2M config.

Arguments

conf::ConfParse: the L2M config object for the task detector.

L2MTaskDetector.TaskDetectorParameters — Type

TaskDetectorParameters

Struct containing the options for the task detector.

Fields

mean::Array{Float64, 1}: means of each feature, used by transformer.

scale::Array{Float64, 1}: scaling parameter of each feature, used by transformer.

transformer::ZScoreTransform: object for transforming the data during preprocessing.

windows::Int: number of cells in each dimension (i.e., windows=2 results in 4 panels).

feat_dim::Int: dimension of each individual feature, before windowing.

opts_ddvfa::opts_DDVFA: the DDVFA options object.

L2MTaskDetector.TaskDetectorParameters — Method

TaskDetectorParameters(config::ConfParse)

Default intialization for the task detector parameters.

Arguments

config::ConfParse: global configuration object with fields for preprocessing.

L2MTaskDetector.classify — Method

classify(taskmod::TaskDetectorModule, x::Array{T, 1}) where {T<:Real}

Get the task from the features.

Arguments

taskmod::TaskDetectorModule: the task detector algorithmic module.

x::Array{Real, 1}: the input sample (assumes preprocessed and 1-D).

L2MTaskDetector.sigmoid — Method

sigmoid(x::Array)

Broadcasts sigmoid(x::Real) across an array x.

Arguments

x::Array: array to process sigmoidally.

L2MTaskDetector.sigmoid — Method

sigmoid(x::Real)

Return the sigmoid function on x.

Arguments

x::Real: the value to process through the sigmoid function.

L2MTaskDetector.task_feature_preprocess — Method

task_feature_preprocess(x::Array)

Preprocess the incoming features for task detection module 3.

Arguments

x::Array: array to preprocess for C3.

params::TaskDetectorParameters: the algorithm's internal parameters.

L2MTaskDetector.train! — Method

train!(taskmod::TaskDetectorModule, x::Array{T, 1} ; y=0)  where {T<:Real}

Train the C3 art module in place on an array.

Arguments

taskmod::TaskDetectorModule: the task detector algorithmic module.

x::Array{Real, 1}: the input sample (assumes preprocessed and 1-D).

y=0: incremental integer supervisory label, default to empty for unsupervised.

L2MTaskDetector.update_metrics — Method

update_metrics(metrics::TaskDetectorMetrics, sample::Array, label::Int)

Update the task detector's metrics using ICVIs.

Fields

metrics::TaskDetectorMetrics: the metrics object being updated.

sample::Array: the array of features that are clustered to the label.

label::Int: the label prescribed by the clustering algorithm.

L2MTaskDetector.window_average — Method

window_average(data::Array, bins::Int)

Average the 3-D data into bins.

For example, if bins = 2, then the features are averaged into two bins per dimension (i.e. four quadrants in a 2-D image), resulting in a vector of length four times the feature length.

Arguments

data::Array: the input features from C1.

bins::Int: the number of cells in each dimension (see above).